logm: Introduce $logm_prefix, and annotate nested guest messages
This allows code elsewhere to annotate log messages which might
otherwise be confusing.  The variable should be localised, and the
value should always start with a space, if not empty.

Use this to annotate the calls to selecthost and selectguest from
within selecthost-for-an-L1.  Otherwise some of the log messages can
be very confusing.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>
Acked-by: Ian Campbell <ian.campbell@citrix.com>