Nested hosts: Use full hostnamepath in various paths
Replace some (but not all) interpolations of $ho->{Name} in filenames
with hostnamepath($ho).

I searched the tree for `{Host}' and `{Name}', looking for paths which
need (or might need) to be qualified with an L0 hostname in an L1 or
L2 case.

I have left ts-logs-capture for now; that will come in a moment.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>
Acked-by: Ian Campbell <ian.campbell@citrix.com>