make-flight: Stop testing ARM for linux-3.14 flights
Since d4dba6183d61 "ts-kernel-build: Include dtbs in dist file" we
know require a "make dtbs_install" target, which was only added after
3.14.

None of the ARM h/w in the XenProject test colo can run with a kernel
this old. We do have a system in the Citrix instance in Cambridge
which does but the complexity vs benefit of testing this doesn't
warrant continuing to test 3.14.

Our current baseline for other ARM testing is a 3.16 based
linux-arm-xen branch.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>
Acked-by: Ian Jackson <Ian.Jackson@eu.citrix.com>