make-flight: consolidate branch_filter_callback for builds and tests
Currently we have a test_matrix_branch_filter_callback which filters
jobs based on the $xenarch and $branch and a separate more adhoc
filter inline for the build jobs.

This has lead to things getting out of sync in the past (e.g. recently
we dropped armhf tests from the linux-3.10 and -3.14 branches but not
the build jobs).

Add a new build_matrix_branch_filter_callback and for make-flight
cause this and test_matrix_branch_filter_callback to use a common
helper.

The adhoc filtering in the build loop remains and will be tidied up
next.

For make-distros-flight just add a nop build filter alongside the test
filter.

No change to the output of standalone-generate-dump-flight-runvars.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>
Acked-by: Ian Jackson <ian.jackson@eu.citrix.com>