Debian: Move sysvinit-core install from preseed_create to preseed_base
preseed_create is used only for physical host installs, but we want
this workaround to apply to at least ts-debian-hvm-install'd guests
which are going to be used for nested testing.

I can't see any harm in doing this globally for all Debian HVM guests,
as well as PV guests installed using Debian Installer.

Signed-off-by: Ian Campbell <ian.campbell@citrix.com>
Acked-by: Ian Jackson <ian.jackson@eu.citrix.com>